[finish]bokkers2

+++ app/controllers/books_controller.rb
@@ -1,6 +1,6 @@
 class BooksController < ApplicationController
   before_action :is_matching_login_user, only: [:edit, :update]
-  
+
   def new
     @book_new = Book.new
   end
@@ -32,23 +32,33 @@ class BooksController < ApplicationController
   end
 
   def edit
+    is_matching_login_user
     @book = Book.find(params[:id])
   end
 
   def update
     @book = Book.find(params[:id])
-    if @book.update(book_params)
-      flash[:notice] = "You have updated book successfully."
-      redirect_to book_path(@book.id)
+    if @book.user != current_user
+      redirect_to books_path
     else
-      render :edit
+      if @book.update(book_params)
+        flash[:notice] = "You have updated book successfully."
+        redirect_to book_path(@book.id)
+      else
+        render :edit
+      end
     end
   end
 
   def destroy
     book = Book.find(params[:id])
-    book.destroy
-    redirect_to books_path
+    if book.user != current_user
+      redirect_to books_path
+    else
+      book.destroy
+      redirect_to books_path
+    end
+
   end
 
 
@@ -57,11 +67,11 @@ class BooksController < ApplicationController
   def book_params
     params.require(:book).permit(:title, :body)
   end
-  
+
   def is_matching_login_user
-    user = User.find(params[:id])
-    unless user.id == current_user.id
-      redirect_to user_path(current_user.id)
+    book = Book.find(params[:id])
+    unless book.user == current_user
+      redirect_to books_path
     end
   end
